Inherit binauraliser functions to binauraliser_nf
